+++ deps/lightrec/lightrec.h
@@ -87,9 +87,9 @@ enum psx_map {
 
 struct lightrec_mem_map_ops {
 	void (*sb)(struct lightrec_state *, u32 opcode,
-		   void *host, u32 addr, u8 data);
+		   void *host, u32 addr, u32 data);
 	void (*sh)(struct lightrec_state *, u32 opcode,
-		   void *host, u32 addr, u16 data);
+		   void *host, u32 addr, u32 data);
 	void (*sw)(struct lightrec_state *, u32 opcode,
 		   void *host, u32 addr, u32 data);
 	u8 (*lb)(struct lightrec_state *, u32 opcode, void *host, u32 addr);
